moved file to the correct location , added all colors for buttons in rgb format
