--- theme/colors.ts
@@ -1,64 +1,0 @@
-const colors = {
-  blue: {
-    50: '#F9FAFB',
-    100: '#F0F2FB',
-    200: '#D9DCE8',
-    300: '#C6C8D3',
-    400: '#B2B3BE',
-    500: '#9D9FA9',
-    600: '#898C95',
-    700: '#63666E',
-    800: '#52545A',
-    900: '#3F444A'
-  },
-  red: {
-    50: '#FCF0ED',
-    100: '#FAE2DC',
-    200: '#F7C3B6',
-    300: '#F4A593',
-    400: '#F7856A',
-    500: '#EB6343',
-    600: '#D13712',
-    700: '#A32B0E',
-    800: '#7F210B',
-    900: '#611908'
-  },
-  yellow: {
-    100: '#FFF8E9',
-    200: '#FFEABA',
-    300: '#FFE099',
-    400: '#FFD26A',
-    500: '#FCBD28',
-    600: '#E0A01F',
-    700: '#BF7D13',
-    800: '#8A570C',
-    900: '#673E0F'
-  },
-  brown: {
-    50: '#F7F5F1',
-    100: '#EDE8DF',
-    200: '#D3CAC0',
-    300: '#B8AEA2',
-    400: '#9F9185',
-    500: '#87756B',
-    600: '#6E5A51',
-    700: '#574139',
-    800: '#412B21',
-    900: '#2D1611'
-  },
-  burgundy: {
-    50: '#F9F3F3',
-    100: '#E6D4D3',
-    200: '#D4B8B4',
-    300: '#C19C96',
-    400: '#AE7F79',
-    500: '#9B655E',
-    600: '#874943',
-    700: '#732E28',
-    800: '#600E0F',
-    900: '#3D0607'
-  },
-  black: '#190d03',
-  white: ''
-};
-export default colors;
